Fixed Bug with Summary and Reasoning
Since AwardPlayer() was/is being called in Start(), everytime the Summary scene was
loaded, the player would be awarded. This could be exploited unintentionally if the
user kept going back and forth between the Reasoning review page and the Summary page.
This was fixed by adding a boolean flag in the static class CaseInformation to track
whether or not awards have been awarded before they are actually given to the user.