Added Experience Level System
The experience level system has been added to the Player class. There are currently four levels the player can
be at any one time. The level number is stored in an integer and managed with PlayerPrefs. There are three
functions that manage the int. One is called from CompletedCase() to update the player's level after each
diagnostic session is completed. The other two are public functions that will return the player's current
level as a number or string.