Fixed bug 10 and messed with shop
The implementation to fix bug 10 has been completed. I just now would like
feedback from the team on the implementation before it is closed out. Additionally,
during a team meeting, I was using this branch to demonstrate the use of scroll views
and attempting to add a vertical bar to the scroll view in Shop. ALL CHANGES TO SHOP
SHOULD BE OVERWRITTEN BY THE BRANCH MAHIR AND HUSAIN ARE WORKING ON.