Fixed Hue or Red Text on Summary
In the summary and the CompareReasoning pages, we use red text to indicate
an incorrect diagnosis or reasoning selection. However, the default red
was quite bright and difficult to look at. As a result, I adjusted the
color to not use the default Unity Color.Red and use a hexidecimal color
to be less jarring.